implementing user roles

+++ app.py
@@ -3,12 +3,12 @@ from flask import Flask,request,jsonify
 from models.user import User
 from database import db
 from flask_login import LoginManager,login_user,current_user,logout_user,login_required
-
+import bcrypt
 
 app = Flask(__name__)
 app.config['SECRET_KEY'] = "your_secret_key"
 BASE_DIR = os.path.abspath(os.path.dirname(__file__))
-app.config['SQLALCHEMY_DATABASE_URI'] = f"sqlite:///{os.path.join(BASE_DIR, 'instance', 'database.db')}"
+app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql+pymysql://root:admin123@127.0.0.1:3306/flask-crud'
 
 
 login_manager = LoginManager()
@@ -29,9 +29,9 @@ def login():
     if username and password:
         user = User.query.filter_by(username=username).first()
 
-        if user and user.password == password:
+        if user and bcrypt.checkpw(str.encode(password),str.encode(user.password)):
                 login_user(user)
-                print(f"{current_user.is_authenticated}autenticado !!")
+                print(f"{current_user.is_authenticated} autenticado !!")
                 return jsonify({"message": "autenticação realizada com sucesso"}), 200
            
     return jsonify({"message": "credenciais invalidas"}), 400
@@ -44,7 +44,6 @@ def logout():
      return jsonify({"message": "logout realizado com sucesso !"})
 
 
-
 @app.route('/user',methods=["POST"])
 def create_user():
      data = request.json
@@ -52,7 +51,8 @@ def create_user():
      password = data.get("password")
 
      if username and password:
-          user = User(username=username, password=password)
+          hashed_password = bcrypt.hashpw(str.encode(password), bcrypt.gensalt())
+          user = User(username=username, password=hashed_password, role='user')
           db.session.add(user)
           db.session.commit()
           return jsonify({"message":"usuario cadastrado com sucesso"})
@@ -75,6 +75,9 @@ def update_user(id_user):
     data = request.json
     user = User.query.get(id_user)
 
+    if id_user != current_user.id and current_user.role == "user":
+         return jsonify ({"message":"operação não permitida"}),403
+    
     if user and data.get("password"):
           user.password = data.get("password")
           db.session.commit()
@@ -82,12 +85,16 @@ def update_user(id_user):
     
     return jsonify({"message":"usuario nao encontrado"}), 404
 
+
 @app.route('/user/<int:id_user>',methods=['DELETE'])
 @login_required
 def delete_user(id_user):
     user = User.query.get(id_user)
     if id_user == current_user.id:
          return jsonify({"message":"delação nao permitida"})
+    
+    if current_user.role != 'admin':
+         return jsonify({"message": "operação nao permitida"}), 403
     if user:
          db.session.delete(user)
          db.session.commit()
